Fix pid_module declaration
pid_moduleの宣言を修正．
引数の型が合わないとコンパイルできなかった．
double型にキャストすることで対応．

+++ Src/cppmain.cpp
@@ -95,8 +95,8 @@ void HAL_TIM_PeriodElapsedCallback(TIM_HandleTypeDef *htim)
 
         // PIDモジュールを作成
         static Pid pid_module[2] = {
-            {motor_setup_data[0].kp,motor_setup_data[0].ki, motor_setup_data[0].kd}, 
-            {motor_setup_data[1].kp,motor_setup_data[1].ki, motor_setup_data[1].kd}
+            {(double)motor_setup_data[0].kp, (double)motor_setup_data[0].ki, (double)motor_setup_data[0].kd},
+            {(double)motor_setup_data[1].kp, (double)motor_setup_data[1].ki, (double)motor_setup_data[1].kd}
         };
 
         // 速度計算
